컴폰전트 분리 진행중
